Modify component to listen for events
When you construct a Component, you specify the types of events
that the component is interested in. This will be defined by the
subclass.

I bet there's a nicer way to do this than by putting it in the
constructer like I'm doing now, though. I don't know what it is,
but I bet it exists.

The test is mostly just to remind myself how unittest works.