Specify event drop-through behaviour
There are 3 actions you can take in a component. You can handle
the event fully (return True), pass the event on unchanged
(return False) or - oh, crabapples. I was going to say re-emit
the events, but what I was thinking was that you should be able
to emit events, as well as pass this event on/not pass it on.

Hmm.