Add test; dispatching across multiple mounts
Test fails, due to the instant error-ing out of an Entity when it
cannot resolve an event. Since you want it to be able to pass
through entities (if you have an attack event, a shield might not
have any reason to evaluate on it) this behaviour should be
changed.