Add a DROPS_THROUGH parameter to Event
This will allow events to be unhandled. For example, a FIRE_ALL
has no single component which will stop it, because it wants to
fire all the things, so it will fully reach the end. Unless you
intentionally put a stopper component at the end it will fall
through, and should.