Push level into event_stack
Basically, for the "DO_ACTION_TO_LEVEL" commands, unless there's
some way of referring to the level in the event handler, you need
to pass the LEVEL down through the event chain.

However, since you have no idea if any action will eventually end
up needing to manipulate the level. For example, A attacking B
might end up needing the level, but it also might not. So where
does that information come in?

Unsure, really; you could put it into the Component that will end
up needing it as internal state, but that skeeves me out,
especially if it's mutable!

So I put it into the event handler.