Raise ValueError if multiple entities on Tile
Well, raise the ValueError if they're on the tile and you attempt
to get the entity on the tile. This will obviously have to
change; the default weapon spawning actually spawns the entity
onto the same tile as the shooter. Self-evidently problematic
here. Also, if projectiles have no collision, obviously they can
stack!

So this is just to surface the issue, not solve it.