Collision events get handler for HANDLER
Instead of attempting to look up the handler from the x, y
coordinates, get the handler from the HANDLER param. This fixes
the ValueError from the 2 entities when firing adjacent to an
enemy.

This will still have a ValueError if you fire into an enemy who
has fired, though. Need to add 'blocking' semantics!