Switch GameStatus from enum->object
This is to allow me to pass it around, and allow things to change
its state.

I feel kinda iffy about this whole thing (passing and setting in
the target object) because it's just so...side-effect-y! But
Python was never meant to be functional. I mean, Guido wanted to
remove lambdas, and there's no tco.

Not to say that Guido's opinions are necessarily Word of God on
what should be done, but...when in Rome? I mean, even then,
though, you might want to avoid the lead plates. I'm not sure
where I'm going with this.