Add public add/remove_component to Entity
Theoretically at some point we want to have some method of
registering component changes to the Level so that you can easily
access them but not unless perf under the new system ends up
degrading.

The old system degraded pretty bad by the end levels but also it
was not optimized literally at all and had horrifying deep-copies
left and right for convenience's sake. So we'll add that in if we
need it.